requirements.txt
streamlit
openai
python-dotenv
requests
pillow